+++ tests/testthat/test-plots.R
@@ -32,3 +32,10 @@ test_that("gg_miss_which_works",{
   vdiffr::expect_doppelganger("gg_miss_which",
                               gg_miss_which_plot)
 })
+
+gg_miss_fct_plot <- gg_miss_fct(riskfactors, marital)
+
+test_that("gg_miss_fct works",{
+  vdiffr::expect_doppelganger("gg_miss_fct",
+                              gg_miss_fct_plot)
+})